- Resolving bugs for SUB and DIV commands in "commands.h"; - Type of numbers written to byte-code was changed from int to special typedef (int32_t) for portable code - Changes in all stack utils typedef for elem_t to int32_t and for stk_size_t to uint32_t

+++ SimpleStack/data_types.h
@@ -14,9 +14,9 @@
 #define CANARY 666
 
 //!@typedef Typedef for values of stack elements
-typedef ptrdiff_t elem_t;
+typedef int32_t elem_t;
 //!@typedef Typedef for sizes params of stack
-typedef size_t stk_size_t;
+typedef uint32_t stk_size_t;
 
 //!@typedef Typedef for Stack_str struct
 typedef struct Stack_str Stack;
